move utilities out of bdf.py
