+++ bolides/bdf.py
@@ -2,22 +2,20 @@ import requests
 from datetime import datetime
 from warnings import warn, filterwarnings
 from tqdm import tqdm
-from math import degrees
 
 import numpy as np
 import pandas as pd
 import pickle
-import ephem
 
 from geopandas import GeoDataFrame
 from cartopy import crs as ccrs
-from shapely.geometry import Point
 
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
 from lightkurve import LightCurve, LightCurveCollection
 
-from . import API_ENDPOINT_EVENTLIST, API_ENDPOINT_EVENT, MPLSTYLE, ROOT_PATH, GLM_FOV_PATH
+from . import API_ENDPOINT_EVENTLIST, API_ENDPOINT_EVENT, MPLSTYLE, ROOT_PATH
+from .utils import make_points
 
 
 class BolideDataFrame(GeoDataFrame):
@@ -90,15 +88,16 @@ class BolideDataFrame(GeoDataFrame):
         super().__init__(init_gdf)
 
         # add additional metadata to the bolides
-        self.annotate_bdf()
+        self.annotate()
 
         from configparser import ConfigParser
         config = ConfigParser()
         config.read(ROOT_PATH+'/desc.ini')
         self.descriptions = config['neo-bolide']
 
-    def annotate_bdf(bdf):
+    def annotate(bdf):
         """Add metadata to bolide detections"""
+        from .astro_utils import get_phase, get_solarhour, get_sun_alt
 
         # lunar phase
         bdf['phase'] = [get_phase(dt) for dt in bdf['datetime']]
@@ -106,18 +105,8 @@ class BolideDataFrame(GeoDataFrame):
         bdf['moon_fullness'] = -np.abs(bdf['phase']-0.5)*2+1
         # solar hour
         bdf['solarhour'] = [get_solarhour(data[0], data[1]) for data in zip(bdf['datetime'], bdf['longitude'])]
-
-        # calculate and add solar altitude
-        sun_alt = []
-        for num, row in bdf.iterrows():
-            obs = ephem.Observer()
-            obs.lon = str(row['longitude'])
-            obs.lat = str(row['latitude'])
-            obs.date = row['datetime']
-            sun = ephem.Sun()
-            sun.compute(obs)
-            sun_alt.append(degrees(sun.alt))
-        bdf['sun_alt'] = sun_alt
+        # solar altitude
+        bdf['sun_alt'] = [get_sun_alt(row) for _, row in bdf.iterrows()]
 
         bdf['date_retrieved'] = datetime.now()
 
@@ -190,6 +179,7 @@ class BolideDataFrame(GeoDataFrame):
         crs = self.geometry.crs
         bdf = self.to_crs(aeqd)
 
+        from .fov_utils import get_boundary
         polygons = [get_boundary(b) for b in boundary]
 
         # either take intersection of FOVs or the union to get a final polygon
@@ -324,6 +314,7 @@ class BolideDataFrame(GeoDataFrame):
                 ax.coastlines()  # plot coastlines
 
             # plot sensor FOV
+            from .fov_utils import add_boundary
             if boundary:
                 add_boundary(ax, boundary, boundary_style)
 
@@ -535,32 +526,12 @@ def get_df_from_usg():
     return gdf
 
 
-def get_feature(feature, bDispObj):
-    return [getattr(disp.features, feature) for disp in bDispObj.bolideDispositionProfileList]
-
-
-# get dict containing a list for every feature of the bDispObj
-def get_features(bDispObj):
-
-    # get a list with a feature dict for each bolide
-    list_of_dicts = [vars(disp.features) for disp in bDispObj.bolideDispositionProfileList]
-
-    # turn it into a dict with a list for every feature
-    # assumption: each dict has the same keys
-    feature_dict = {key: [dic[key] for dic in list_of_dicts] for key in list_of_dicts[0]}
-
-    return feature_dict
-
-
-# wrapper class for loading from pickled pipeline data
-class Wrapper():
-    def __init__(self):
-        pass
-
-
 def get_df_from_pipeline(files, use_pickle=False):
     from bolide_dispositions import BolideDispositions as bdisp
+    from .pipeline_utils import get_features
+
     if use_pickle:
+        from .utils import Wrapper
         with open(files, 'rb') as f:
             bdisplist = pickle.load(f)
             bDispObj = Wrapper()
@@ -591,132 +562,6 @@ def get_df_from_pipeline(files, use_pickle=False):
     bdf = bdf.rename(columns=column_translation)
 
     return bdf
-
-
-def make_points(lons, lats):
-    coords = zip(lons, lats)
-    points = [Point(coord[0], coord[1]) for coord in coords]
-    return points
-
-
-def get_phase(datetime):
-    date = ephem.Date(datetime)
-    nnm = ephem.next_new_moon(date)
-    pnm = ephem.previous_new_moon(date)
-
-    lunation = (date-pnm)/(nnm-pnm)
-    return lunation
-
-
-def get_solarhour(datetime, lon):
-    o = ephem.Observer()
-    o.date = datetime
-    from math import pi
-    o.long = lon/180 * pi
-    sun = ephem.Sun()
-    sun.compute(o)
-    hour_angle = o.sidereal_time() - sun.ra
-    solarhour = ephem.hours(hour_angle+ephem.hours('12:00')).norm/(2*pi) * 24
-    return solarhour
-
-
-def add_boundary(ax, boundary=None, boundary_style={}):
-
-    assert boundary is not None
-
-    # for filename in boundary:
-    #     with open('data/'+filename+'.pkl', 'rb') as f:
-    #         b = pickle.load(f)
-    #     ax.add_geometries([b[0]], crs=b[1], facecolor='none',
-    #                       edgecolor='k', alpha=1, linewidth=3)
-
-    boundary_defaults = {"facecolor": "none"}
-    for key, value in boundary_defaults.items():
-        if key not in boundary_style:
-            boundary_style[key] = value
-
-    polygons = [get_boundary(b) for b in boundary]
-
-    crs = ccrs.AzimuthalEquidistant(central_latitude=90)
-    ax.add_geometries(polygons, crs=crs, **boundary_style)
-
-
-def get_boundary(boundary):
-    from netCDF4 import Dataset
-    from shapely.geometry import Polygon
-
-    if boundary == 'goes-w-ni':
-        fov = Dataset(GLM_FOV_PATH, "r", format="NETCDF4")
-        lats = fov.variables['G17_fov_lat'][0]
-        lons = fov.variables['G17_fov_lon'][0]
-        polygon = Polygon(zip(lons, lats))
-
-    elif boundary == 'goes-w-i':
-        fov = Dataset(GLM_FOV_PATH, "r", format="NETCDF4")
-        lats = fov.variables['G17_fov_lat_inverted'][0]
-        lons = fov.variables['G17_fov_lon_inverted'][0]
-        polygon = Polygon(zip(lons, lats))
-
-    elif boundary == 'goes-17-89.5':
-        fov = Dataset(GLM_FOV_PATH, "r", format="NETCDF4")
-        lats = fov.variables['G17_fov_lat'][0]
-        lons = np.array(fov.variables['G17_fov_lon'][0]) + (-89.5-(-137.2))
-        polygon = Polygon(zip(lons, lats))
-
-    elif boundary == 'goes-w':
-        from shapely.ops import unary_union
-
-        # get data and create polygons
-        fov = Dataset(GLM_FOV_PATH, "r", format="NETCDF4")
-        goes_w_ni = get_boundary('goes-w-ni')
-        goes_w_i = get_boundary('goes-w-i')
-        return unary_union([goes_w_ni, goes_w_i])
-
-    elif boundary == 'goes-e':
-        fov = Dataset(GLM_FOV_PATH, "r", format="NETCDF4")
-        lats = fov.variables['G16_fov_lat'][0]
-        lons = fov.variables['G16_fov_lon'][0]
-        polygon = Polygon(zip(lons, lats))
-
-    elif boundary == 'fy4a-n':
-        lons = [55, 157.4, 127, 81.75]
-        lats = [56.25, 56.25, 14.92, 14.92]
-        polygon = fy4a_corners_to_boundary(lons, lats)
-
-    elif boundary == 'fy4a-s':
-        lons = [81.75, 127, 157.4, 55]
-        lats = [-14.92, -14.92, -56.25, -56.25]
-        polygon = fy4a_corners_to_boundary(lons, lats)
-
-    return aeqd_from_lonlat(polygon)
-
-
-def aeqd_from_lonlat(polygon):
-    import pyproj
-    # define Azimuthal Equidistant projection
-    aeqd = pyproj.Proj(proj='aeqd', ellps='WGS84', datum='WGS84', lat_0=90, lon_0=0).srs
-    gdf = GeoDataFrame(geometry=[polygon], crs='epsg:4236')
-    gdf = gdf.to_crs(aeqd)
-    return gdf.geometry[0]
-
-
-# get corner points of fy4a FOV and return boundary
-def fy4a_corners_to_boundary(lons, lats):
-    import pyproj
-    from shapely.geometry import Point, Polygon, LinearRing
-
-    points = [Point(lon, lat) for lon, lat in zip(lons, lats)]
-    gdf = GeoDataFrame(geometry=points, crs='epsg:4236')
-    geos = pyproj.Proj(proj='geos', ellps='WGS84', datum='WGS84', h=35785831.0, lon_0=105).srs
-    gdf = gdf.to_crs(geos)
-    points = gdf.geometry
-    linestring = LinearRing(zip([p.x for p in points], [p.y for p in points]))
-    length = linestring.length
-    points = [linestring.interpolate(x) for x in np.linspace(0, length, 1000)]
-    polygon = Polygon(zip([p.x for p in points], [p.y for p in points]))
-    gdf = GeoDataFrame(geometry=[polygon], crs=geos)
-    gdf = gdf.to_crs('epsg:4236')
-    return gdf.geometry[0]
 
 
 def force_bdf_class(bdf):
